Adjust metric name validation unit test
The metric name validation unit test had to be adjusted as the
validation in the Prometheus Go Client Library was
[changed](https://github.com/prometheus/client_golang/releases/tag/v1.21.0) and now all
the UTF8 characters are valid in the name of the metrics.

The name of the metric in the test was adjusted as per [this](https://github.com/prometheus/common/blob/main/model/metric_test.go#L144) unit test in the Prometheus common repo.